@@ -12,102 +17,120 @@
   "dependencies": {
     "@zeit/next-css": "^1.0.1",
     "@zeit/next-sass": "^1.0.1",
+    "animejs": "^3.1.0",
     "array-to-sentence": "^2.0.0",
-    "axios": "^0.18.0",
+    "axios": "^0.19.2",
     "balloon-css": "^0.4.0",
-    "canvas": "^2.3.1",
+    "canvas": "^2.6.1",
     "classnames": "^2.2.6",
     "debounce": "^1.2.0",
     "debug": "^4.1.1",
-    "dompurify": "^1.0.9",
-    "dotenv": "^4.0.0",
-    "dotenv-defaults": "^1.0.2",
-    "execa": "^1.0.0",
-    "fabric": "^2.4.3",
-    "firebase": "^5.8.0",
-    "flatten": "^1.0.2",
-    "git-url-parse": "^10.1.0",
+    "dompurify": "^2.0.8",
+    "dotenv": "^8.2.0",
+    "dotenv-defaults": "^1.1.1",
+    "execa": "^4.0.0",
+    "fabric": "^3.6.3",
+    "firebase": "^6.6.0",
+    "flatten": "^1.0.3",
+    "git-url-parse": "^11.1.2",
     "github": "^10.0.0",
     "got": "^9.6.0",
-    "hot-shots": "^6.1.1",
+    "hot-shots": "^6.8.5",
     "ipchecker": "^0.0.2",
     "is-empty-object": "^1.1.1",
-    "jsdom": "^13.1.0",
-    "koa": "^2.6.1",
-    "koa-basic-auth": "^3.0.0",
+    "jsdom": "^16.2.1",
+    "koa": "^2.11.0",
+    "koa-basic-auth": "^4.0.0",
     "koa-better-ratelimit": "^2.1.2",
-    "koa-bodyparser": "^4.2.1",
+    "koa-bodyparser": "^4.3.0",
     "koa-cache-control": "^2.0.0",
     "koa-cash": "3.0.1-0",
     "koa-compress": "^3.0.0",
     "koa-proxy": "^0.9.0",
-    "koa-requestid": "^2.0.1",
+    "koa-requestid": "^2.1.0",
     "koa-router": "^7.4.0",
     "koa-send": "^5.0.0",
     "koa-static": "^5.0.0",
     "lodash.isequal": "^4.5.0",
     "lodash.sortby": "^4.7.0",
-    "lru-cache": "^4.1.3",
-    "memory-fs": "^0.4.1",
+    "lru-cache": "^5.1.1",
+    "memory-fs": "^0.5.0",
     "mkdir-promise": "^1.0.0",
     "natural": "^0.6.3",
-    "next": "7.0.2",
-    "node-fetch": "^2.2.0",
+    "next": "^8.1.0",
+    "node-fetch": "^2.6.0",
     "normalize.css": "^7.0.0",
     "p-queue": "^3.0.0",
-    "package-build-stats": "4.1.0",
+    "package-build-stats": "6.2.1",
+    "pacote": "^11.1.4",
     "performance-now": "^2.1.0",
-    "promise-queue-plus": "^1.2.1",
+    "promise-queue-plus": "^1.2.2",
     "promise.series": "^0.2.0",
-    "prop-types": "^15.6.2",
-    "query-string": "^6.2.0",
-    "react": "16.7.0",
+    "prop-types": "^15.7.2",
+    "query-string": "^6.11.1",
+    "react": "16.13.1",
     "react-autocomplete": "^1.8.1",
-    "react-dom": "16.7.0",
+    "react-dom": "16.13.1",
     "react-dropzone": "^7.0.1",
-    "react-flip-move": "^3.0.3",
-    "react-ga": "^2.5.6",
+    "react-flip-move": "^3.0.4",
+    "react-ga": "^2.7.0",
     "react-sidebar": "^3.0.2",
     "remark": "^9.0.0",
-    "semver": "^5.6.0",
-    "strip-markdown": "^3.0.2",
+    "semver": "^7.1.3",
+    "strip-markdown": "^3.1.2",
     "sumo-logger": "^1.6.0",
     "trending-github": "^1.0.2",
-    "uglifyjs-webpack-plugin": "^2.1.1",
-    "unfetch": "^4.0.1",
-    "webpack": "^4.28.4",
-    "winston": "^3.1.0",
+    "uglifyjs-webpack-plugin": "^2.2.0",
+    "unfetch": "^4.1.0",
+    "webpack": "^4.42.1",
+    "winston": "^3.2.1",
+    "winston-graylog2": "^2.1.2",
     "winston-logzio": "^3.0.3",
     "winston-sumologic-transport": "^1.1.0",
     "workerpool": "^3.0.0"
   },
   "scripts": {
     "dev": "DEBUG=bp:* node index",
-    "build": "next build",
+    "build": "NODE_ENV=production next build",
     "prestart": "yarn build",
     "start": "NODE_ENV=production DEBUG=bp:* node index",
     "prod": "yarn run build && yarn start",
     "test": "jest",
     "test:watch": "jest --watch",
     "postinstall": "yarn run build",
+    "prettier": "prettier --write '**/*.{html,js,json,css,scss,jsx,flow,md,yml,yaml}'",
     "deploy": "git branch -D gh-pages && git checkout -b gh-pages && yarn run build && yarn run export && cp -R out/* . "
   },
+  "husky": {
+    "hooks": {
+      "pre-commit": "pretty-quick --staged"
+    }
+  },
+  "prettier": {
+    "semi": false,
+    "arrowParens": "avoid",
+    "singleQuote": true
+  },
   "devDependencies": {
-    "autoprefixer": "^7.1.2",
-    "babel-jest": "^20.0.3",
+    "@babel/core": "^7.9.0",
+    "@babel/preset-env": "^7.9.0",
+    "@zeit/next-source-maps": "^0.0.4-canary.1",
+    "autoprefixer": "^9.7.5",
+    "babel-jest": "^24.9.0",
     "babel-plugin-css-modules-transform": "^1.6.1",
-    "babel-plugin-inline-react-svg": "^1.0.0",
-    "babel-plugin-module-resolver": "^3.1.1",
-    "babel-plugin-transform-optional-chaining": "^7.0.0-beta.3",
+    "babel-plugin-inline-react-svg": "^1.1.1",
+    "babel-plugin-module-resolver": "^4.0.0",
     "babel-plugin-wrap-in-js": "^1.1.1",
     "css-loader": "^1.0.1",
-    "glob": "^7.1.2",
-    "jest": "^20.0.4",
-    "node-sass": "^4.11.0",
-    "postcss": "^7.0.13",
+    "glob": "^7.1.6",
+    "husky": "^4.2.3",
+    "jest": "^24.9.0",
+    "node-sass": "^4.13.1",
+    "postcss": "^7.0.27",
     "postcss-easy-import": "^3.0.0",
     "postcss-loader": "^3.0.0",
-    "raw-loader": "^0.5.1",
-    "sass-loader": "^7.1.0"
+    "prettier": "2.0.2",
+    "pretty-quick": "^2.0.1",
+    "sass-loader": "^8.0.2"
   }
 }
