@@ -1,18 +1,19 @@
-import { withRouter } from "next/router";
+import Router, { withRouter } from 'next/router'
 import React, { Component } from 'react'
 import Analytics from 'react-ga'
-import FlipMove from 'react-flip-move';
+import FlipMove from 'react-flip-move'
 import cx from 'classnames'
 
-const PromiseQueue = require('p-queue');
-import Stat from '../result/Stat'
+const PromiseQueue = require('p-queue')
+const queryString = require('query-string')
+import Stat from 'client/components/Stat'
 import Link from 'next/link'
 import ResultLayout from '../../client/components/ResultLayout'
 import { parsePackageString } from 'utils/common.utils'
 
 import './ScanResults.scss'
 import API from 'client/api'
-import {getTimeFromSize} from 'utils'
+import { getTimeFromSize } from 'utils'
 
 class ResultCard extends Component {
   render() {
@@ -23,9 +24,7 @@ class ResultCard extends Component {
     switch (pack.promiseState) {
       case 'pending':
         content = (
-          <div className="scan-results__loading-text">
-            Calculating &hellip;
-          </div>
+          <div className="scan-results__loading-text">Calculating &hellip;</div>
         )
         break
 
@@ -34,29 +33,29 @@ class ResultCard extends Component {
           <div className="scan-results__stat-container">
             <Stat
               className="scan-results__stat-item"
-              value={ pack.result.size }
-              type={ Stat.type.SIZE }
+              value={pack.result.size}
+              type={Stat.type.SIZE}
               label="Min"
               compact
             />
             <Stat
               className="scan-results__stat-item"
-              value={ pack.result.gzip }
-              type={ Stat.type.SIZE }
+              value={pack.result.gzip}
+              type={Stat.type.SIZE}
               label="Min + GZIP"
               compact
             />
             <Stat
               className="scan-results__stat-item"
-              value={ getTimeFromSize(pack.result.gzip).twoG }
-              type={ Stat.type.TIME }
+              value={getTimeFromSize(pack.result.gzip).twoG}
+              type={Stat.type.TIME}
               label="2G Edge"
               compact
             />
             <Stat
               className="scan-results__stat-item"
-              value={ getTimeFromSize(pack.result.gzip).threeG }
-              type={ Stat.type.TIME }
+              value={getTimeFromSize(pack.result.gzip).threeG}
+              type={Stat.type.TIME}
               label="Emerging 3G"
               compact
             />
@@ -67,38 +66,38 @@ class ResultCard extends Component {
       case 'rejected':
         content = (
           <details className="scan-results__error-text">
-            <summary> { pack.error.code }</summary>
-            <p dangerouslySetInnerHTML={ { __html: pack.error.message } } />
+            <summary> {pack.error.code}</summary>
+            <p dangerouslySetInnerHTML={{ __html: pack.error.message }} />
           </details>
         )
         break
     }
 
     return (
-      <li className={ cx('scan-results__item', {
+      <li
+        className={cx('scan-results__item', {
           'scan-results__item--loading': pack.promiseState === 'pending',
           'scan-results__item--error': pack.promiseState === 'rejected',
-        },
-      ) }>
+        })}
+      >
         <div className="scan-results__index">
-          <span> { index + 1 }. </span>
+          <span> {index + 1}. </span>
         </div>
-        <div className="scan-results__name" data-name={ pack.name }>
-          <Link href={ `/result?p=${pack.packageString}` }>
+        <div className="scan-results__name" data-name={pack.name}>
+          <Link href={`/result?p=${pack.packageString}`}>
             <a>
-              <span className="scan-results__package-name"> { pack.name }</span>
+              <span className="scan-results__package-name"> {pack.name}</span>
               <div>
-                { pack.version && (
+                {pack.version && (
                   <span className="scan-results__package-version">
-                    v{ pack.version }
-                    </span>
-                ) }
+                    v{pack.version}
+                  </span>
+                )}
               </div>
             </a>
           </Link>
-
         </div>
-        { content }
+        {content}
       </li>
     )
   }
@@ -109,8 +108,10 @@ class ScanResults extends Component {
     super(props)
 
     const { router } = this.props
+    const sortMode = router.query.sortMode
     const packageStrings = router.query.packages
-    const packages = packageStrings.split(',')
+    const packages = packageStrings
+      .split(',')
       .map(str => str.trim())
       .map(str => ({
         promiseState: 'pending',
@@ -118,50 +119,54 @@ class ScanResults extends Component {
         ...parsePackageString(str),
       }))
 
-    this.state = { packages, sortMode: null }
+    this.state = { packages, sortMode: sortMode }
   }
 
   componentDidMount() {
     const { packages } = this.state
-    const queue = new PromiseQueue({ concurrency: 3 });
+    const queue = new PromiseQueue({ concurrency: 3 })
     const startTime = Date.now()
 
     packages.forEach(pack => {
-      queue.add(() => API.getInfo(pack.packageString)
-        .then((result) => {
-          this.updatePackageState(pack, {
-            promiseState: 'fulfilled',
-            version: result.version,
-            result,
-          })
+      queue.add(() =>
+        API.getInfo(pack.packageString)
+          .then(result => {
+            this.updatePackageState(pack, {
+              promiseState: 'fulfilled',
+              version: result.version,
+              result,
+            })
 
-          Analytics.event({
-            category: 'Search',
-            action: 'Search Success',
-            label: pack.packageString.replace(/@/g, '[at]'),
+            Analytics.event({
+              category: 'Search',
+              action: 'Search Success',
+              label: pack.packageString.replace(/@/g, '[at]'),
+            })
           })
-        })
-        .catch(({ error }) => {
-          console.error(error)
-          this.updatePackageState(pack, {
-            promiseState: 'rejected',
-            error,
+          .catch(({ error }) => {
+            console.error(error)
+            this.updatePackageState(pack, {
+              promiseState: 'rejected',
+              error,
+            })
+            Analytics.event({
+              category: 'Scan',
+              action: 'Search Failure',
+              label: pack.packageString.replace(/@/g, '[at]'),
+            })
           })
-          Analytics.event({
-            category: 'Scan',
-            action: 'Search Failure',
-            label: pack.packageString.replace(/@/g, '[at]'),
-          })
-        }))
+      )
     })
 
     queue.onIdle().then(() => {
-      const successfulBuildCount = packages.reduce((curSum, nextPack) =>
-          nextPack.promiseState === 'fulfilled' ? curSum + 1 : curSum
-        , 0)
+      const successfulBuildCount = packages.reduce(
+        (curSum, nextPack) =>
+          nextPack.promiseState === 'fulfilled' ? curSum + 1 : curSum,
+        0
+      )
 
-      Analytics.set({ metric1: packages.length });
-      Analytics.set({ metric2: successfulBuildCount / packages.length });
+      Analytics.set({ metric1: packages.length })
+      Analytics.set({ metric2: successfulBuildCount / packages.length })
 
       Analytics.event({
         category: 'scan',
@@ -172,7 +177,7 @@ class ScanResults extends Component {
         category: 'scan',
         variable: 'scan results',
         value: Date.now() - startTime,
-      });
+      })
     })
 
     Analytics.pageview(window.location.pathname)
@@ -180,8 +185,8 @@ class ScanResults extends Component {
 
   updatePackageState(pack, state) {
     const { packages } = this.state
-    const packIndex = packages.findIndex(({ packageString }) =>
-      packageString === pack.packageString,
+    const packIndex = packages.findIndex(
+      ({ packageString }) => packageString === pack.packageString
     )
 
     packages[packIndex] = {
@@ -192,101 +197,112 @@ class ScanResults extends Component {
     this.setState({ packages })
   }
 
-  handleSortAlphabetic = () => {
-    const { packages } = this.state
-    const sortedList = packages.sort((packA, packB) =>
-      packA.name.localeCompare(packB.name),
-    )
+  setParamsAndState = ( sortMode ) => {
+    const updatedQuery = { ...this.props.router.query, sortMode}
+    Router.replace(`/scan-results?${queryString.stringify(updatedQuery,{encode: false})}`)
 
-    this.setState({ packages: sortedList, sortMode: 'alphabetic' })
+    this.setState({ sortMode: sortMode })
+  }
+
+  handleSortAlphabetic = () => {
+    this.setParamsAndState('alphabetic')
   }
 
   handleSortSize = () => {
-    const { packages } = this.state
-    const sortedList = packages.sort((packA, packB) => {
-      const packASize = packA.result ? packA.result.gzip : 0
-      const packBSize = packB.result ? packB.result.gzip : 0
+    this.setParamsAndState('size')
+  }
 
-      return packBSize - packASize
-    })
+  sortPackages = () => {
+    const { packages, sortMode } = this.state
+    let sortedList
 
-    this.setState({ packages: sortedList, sortMode: 'size' })
+    if (sortMode === 'size') {
+      sortedList = packages.sort((packA, packB) => {
+        const packASize = packA.result ? packA.result.gzip : 0
+        const packBSize = packB.result ? packB.result.gzip : 0
+
+        return packBSize - packASize
+      })
+    } else {
+      sortedList = packages.sort((packA, packB) =>
+        packA.name.localeCompare(packB.name)
+      )
+    }
+    return sortedList
   }
 
   render() {
-    const { packages, sortMode } = this.state
+    const { sortMode } = this.state
+    const packages = this.sortPackages()
+
     const totalMinSize = packages
       .reduce((curTotal, pack) =>
         curTotal + (pack.result ? pack.result.size : 0), 0)
 
-    const totalGZIPSize = packages
-      .reduce((curTotal, pack) => curTotal +
-        (pack.result ? pack.result.gzip : 0), 0)
+    const totalGZIPSize = packages.reduce(
+      (curTotal, pack) => curTotal + (pack.result ? pack.result.gzip : 0),
+      0
+    )
 
     return (
       <ResultLayout className="scan-results">
-
         <h1> Results</h1>
         <div className="scan-results__sort-panel">
           <label> Sort By: </label>
           <button
-            className={ cx({
+            className={cx({
               'scan-results__sort--selected': sortMode === 'alphabetic',
-            }) }
-            onClick={ this.handleSortAlphabetic }
+            })}
+            onClick={this.handleSortAlphabetic}
           >
             Name: A &rarr; Z
           </button>
           <button
-            className={ cx({
+            className={cx({
               'scan-results__sort--selected': sortMode === 'size',
-            }) }
-            onClick={ this.handleSortSize }
+            })}
+            onClick={this.handleSortSize}
           >
             Size: High &rarr; Low
           </button>
         </div>
         <ul className="scan-results__container">
-          <FlipMove duration={ 350 } easing="cubic-bezier(0.175, 0.885, 0.325, 1.040)">
-            { packages
-              .map((pack, index) => (
-                <ResultCard
-                  pack={ pack }
-                  index={ index }
-                  key={ pack.name }
-                />
-              )) }
+          <FlipMove
+            duration={350}
+            easing="cubic-bezier(0.175, 0.885, 0.325, 1.040)"
+          >
+            {packages.map((pack, index) => (
+              <ResultCard pack={pack} index={index} key={pack.name} />
+            ))}
           </FlipMove>
           <li className="scan-results__item scan-results__item--total">
-            <div className="scan-results__name">
-              Total
-            </div>
+            <div className="scan-results__name">Total</div>
             <div className="scan-results__stat-container">
               <Stat
                 className="scan-results__stat-item"
-                value={ totalMinSize }
-                type={ Stat.type.SIZE }
+                value={totalMinSize}
+                type={Stat.type.SIZE}
                 label="Min"
                 compact
               />
               <Stat
                 className="scan-results__stat-item"
-                value={ totalGZIPSize }
-                type={ Stat.type.SIZE }
+                value={totalGZIPSize}
+                type={Stat.type.SIZE}
                 label="Min + GZIP"
                 compact
               />
               <Stat
                 className="scan-results__stat-item"
-                value={ totalGZIPSize / 1024 / 30 }
-                type={ Stat.type.TIME }
+                value={totalGZIPSize / 1024 / 30}
+                type={Stat.type.TIME}
                 label="2G Edge"
                 compact
               />
               <Stat
                 className="scan-results__stat-item"
-                value={ totalGZIPSize / 1024 / 50 }
-                type={ Stat.type.TIME }
+                value={totalGZIPSize / 1024 / 50}
+                type={Stat.type.TIME}
                 label="Emerging 3G"
                 compact
               />
