@@ -6,40 +6,47 @@ export default class API {
       Accept: 'application/json',
     }
 
-    if(isInternal) {
-      headers['X-Bundlephobia-User'] =  'bundlephobia website'
+    if (isInternal) {
+      headers['X-Bundlephobia-User'] = 'bundlephobia website'
     }
-    return fetch(url, { headers })
-      .then(res => {
-        if (!res.ok) {
-          try {
-            return res.json()
-              .then(err => Promise.reject(err))
-          } catch (e) {
-            // Heroku might have timed out / shut down
-            if (res.status === 503) {
-              return Promise.reject({
-                error: {
-                  code: 'TimeoutError',
-                  message: 'This is taking unusually long. Check back in a couple of minutes?',
-                },
-              })
-            }
-
+    return fetch(url, { headers }).then(res => {
+      if (!res.ok) {
+        try {
+          return res.json().then(err => Promise.reject(err))
+        } catch (e) {
+          if (res.status === 503) {
             return Promise.reject({
               error: {
-                code: 'BuildError',
-                message: 'Oops, something went wrong and we don\'t have an appropriate error for this. Open an issue maybe?'
-              }
+                code: 'TimeoutError',
+                message:
+                  'This is taking unusually long. Check back in a couple of minutes?',
+              },
             })
           }
+
+          return Promise.reject({
+            error: {
+              code: 'BuildError',
+              message:
+                "Oops, something went wrong and we don't have an appropriate error for this. Open an issue maybe?",
+            },
+          })
         }
-        return res.json()
-      })
+      }
+      return res.json()
+    })
   }
 
   static getInfo(packageString) {
     return API.get(`/api/size?package=${packageString}&record=true`)
+  }
+
+  static getExports(packageString) {
+    return API.get(`/api/exports?package=${packageString}`)
+  }
+
+  static getExportsSizes(packageString) {
+    return API.get(`/api/exports-sizes?package=${packageString}`)
   }
 
   static getHistory(packageString) {
@@ -55,27 +62,26 @@ export default class API {
   }
 
   static getSuggestions(query) {
-
     const suggestionSort = (packageA, packageB) => {
       // Rank closely matching packages followed
       // by most popular ones
       if (
         Math.abs(
-          Math.log(packageB.searchScore) -
-          Math.log(packageA.searchScore),
+          Math.log(packageB.searchScore) - Math.log(packageA.searchScore)
         ) > 1
       ) {
         return packageB.searchScore - packageA.searchScore
       } else {
-        return packageB.score.detail.popularity -
-          packageA.score.detail.popularity
+        return (
+          packageB.score.detail.popularity - packageA.score.detail.popularity
+        )
       }
     }
 
-
-    return API
-      .get(`https://api.npms.io/v2/search/suggestions?q=${query}`, false)
-      .then(result => result.sort(suggestionSort))
+    return API.get(
+      `https://api.npms.io/v2/search/suggestions?q=${query}`,
+      false
+    ).then(result => result.sort(suggestionSort))
 
     //backup when npms.io is down
 
@@ -84,7 +90,7 @@ export default class API {
     //    .sort(suggestionSort)
     //    .map(suggestion => {
     //      const name = suggestion.package.name
-    //      const hasMatch = name.indexOf(query) > -1
+    //      const hasMatch = name.includes(query)
     //      const startIndex = name.indexOf(query)
     //      const endIndex = startIndex + query.length
     //      let highlight
