@@ -1,14 +1,15 @@
 import React, { PureComponent } from 'react'
-import API from 'client/api'
+import debounce from 'debounce'
 import AutoComplete from 'react-autocomplete'
 import cx from 'classnames'
-import './AutocompleteInput.scss'
-import debounce from 'debounce'
 
+import SearchIcon from '../Icons/SearchIcon'
 import { parsePackageString } from 'utils/common.utils'
+import API from 'client/api'
+
+import './AutocompleteInput.scss'
 
 export default class AutocompleteInput extends PureComponent {
-
   static defaultProps = {
     initialValue: '',
   }
@@ -19,26 +20,22 @@ export default class AutocompleteInput extends PureComponent {
     isMenuVisible: false,
   }
 
-  getSuggestions = debounce(
-    value => {
-      API.getSuggestions(value)
-        .then(result => {
-          this.setState({ suggestions: result })
-        })
-    },
-    150,
-  )
+  getSuggestions = debounce(value => {
+    API.getSuggestions(value).then(result => {
+      this.setState({ suggestions: result })
+    })
+  }, 150)
 
   renderSuggestionItem = (item, isHighlighted) => (
     <div
-      className={ cx('autocomplete-input__suggestion', {
+      className={cx('autocomplete-input__suggestion', {
         'autocomplete-input__suggestion--highlight': isHighlighted,
-      }) }
+      })}
     >
-      <div dangerouslySetInnerHTML={ { __html: item.highlight } } />
+      <div dangerouslySetInnerHTML={{ __html: item.highlight }} />
 
       <div className="autocomplete-input__suggestion-description">
-        { item.package.description }
+        {item.package.description}
       </div>
     </div>
   )
@@ -47,7 +44,7 @@ export default class AutocompleteInput extends PureComponent {
     const { onSearchSubmit } = this.props
 
     if (e) {
-      e.preventDefault();
+      e.preventDefault()
     }
 
     onSearchSubmit(value || this.state.value)
@@ -63,7 +60,7 @@ export default class AutocompleteInput extends PureComponent {
     }
   }
 
-  handleMenuVisibilityChange = (isOpen) => {
+  handleMenuVisibilityChange = isOpen => {
     this.setState({ isMenuVisible: isOpen })
   }
 
@@ -71,24 +68,29 @@ export default class AutocompleteInput extends PureComponent {
     const { className, containerClass, autoFocus } = this.props
     const { suggestions, value, isMenuVisible } = this.state
     const { name, version } = parsePackageString(value)
-    const baseFontSize = (typeof window !== 'undefined' && window.innerWidth < 640) ?
-      22 : 35
-    const maxFullSizeChars = (typeof window !== 'undefined' && window.innerWidth < 640) ?
-      15 : 20
-    const searchFontSize = value.length < maxFullSizeChars ? null :
-      `${baseFontSize - (value.length - maxFullSizeChars) * 0.8}px`
+    const baseFontSize =
+      typeof window !== 'undefined' && window.innerWidth < 640 ? 22 : 35
+    const maxFullSizeChars =
+      typeof window !== 'undefined' && window.innerWidth < 640 ? 15 : 20
+    const searchFontSize =
+      value.length < maxFullSizeChars
+        ? null
+        : `${baseFontSize - (value.length - maxFullSizeChars) * 0.8}px`
 
     return (
       <form
-        className={ cx(containerClass, "autocomplete-input__form") }
-        onSubmit={ this.handleSubmit }
+        className={cx(containerClass, 'autocomplete-input__form')}
+        onSubmit={this.handleSubmit}
       >
-        <div className={ cx("autocomplete-input__container", className, {
-          "autocomplete-input__container--menu-visible": isMenuVisible && !!suggestions.length
-        }) }>
+        <div
+          className={cx('autocomplete-input__container', className, {
+            'autocomplete-input__container--menu-visible':
+              isMenuVisible && !!suggestions.length,
+          })}
+        >
           <AutoComplete
-            getItemValue={ item => item.package.name }
-            inputProps={ {
+            getItemValue={item => item.package.name}
+            inputProps={{
               placeholder: 'find package',
               className: 'autocomplete-input',
               autoCorrect: 'off',
@@ -96,63 +98,48 @@ export default class AutocompleteInput extends PureComponent {
               autoCapitalize: 'off',
               spellCheck: false,
               style: { fontSize: searchFontSize },
-            } }
-            onMenuVisibilityChange={ this.handleMenuVisibilityChange }
-            onChange={ this.handleInputChange }
-            ref={ s => this.searchInput = s }
-            value={ value }
-            items={ suggestions }
-            onSelect={ (value, item) => {
+            }}
+            onMenuVisibilityChange={this.handleMenuVisibilityChange}
+            onChange={this.handleInputChange}
+            ref={s => (this.searchInput = s)}
+            value={value}
+            items={suggestions}
+            onSelect={(value, item) => {
               this.setState({ value, suggestions: [item] })
               this.handleSubmit(null, null, value)
-            } }
-            renderMenu={
-              (items, value, inbuiltStyles) => {
-                return (
-                  <div
-                    style={ { minWidth: inbuiltStyles.minWidth } }
-                    className="autocomplete-input__suggestions-menu"
-                    children={ items }
-                  />
-                )
-              }
-            }
-            wrapperStyle={ {
+            }}
+            renderMenu={(items, value, inbuiltStyles) => {
+              return (
+                <div
+                  style={{ minWidth: inbuiltStyles.minWidth }}
+                  className="autocomplete-input__suggestions-menu"
+                  children={items}
+                />
+              )
+            }}
+            wrapperStyle={{
               display: 'inline-block',
               width: '100%',
               position: 'relative',
-            } }
-            renderItem={ this.renderSuggestionItem }
+            }}
+            renderItem={this.renderSuggestionItem}
           />
           <div
-            style={ { fontSize: searchFontSize } }
+            style={{ fontSize: searchFontSize }}
             className="autocomplete-input__dummy-input"
           >
-      <span className="dummy-input__package-name">
-       { name }
-      </span>
-            {
-              version !== null && (
-                <span className="dummy-input__at-separator">
-          @
-        </span>
-              )
-            }
-            <span className="dummy-input__package-version">
-        { version }
-      </span>
+            <span className="dummy-input__package-name">{name}</span>
+            {version !== null && (
+              <span className="dummy-input__at-separator">@</span>
+            )}
+            <span className="dummy-input__package-version">{version}</span>
           </div>
         </div>
-        <div className="autocomplete-input__search-icon"
-             onClick={ this.handleSubmit }>
-          <svg
-            width="90"
-            height="90"
-            viewBox="0 0 90 90"
-            xmlns="http://www.w3.org/2000/svg"
-          >
-            <path d="M89.32 86.5L64.25 61.4C77.2 47 76.75 24.72 62.87 10.87 55.93 3.92 46.7.1 36.87.1s-19.06 3.82-26 10.77C3.92 17.8.1 27.05.1 36.87s3.82 19.06 10.77 26c6.94 6.95 16.18 10.77 26 10.77 9.15 0 17.8-3.32 24.55-9.4l25.08 25.1c.38.4.9.57 1.4.57.52 0 1.03-.2 1.42-.56.78-.78.78-2.05 0-2.83zM36.87 69.63c-8.75 0-16.98-3.4-23.17-9.6-6.2-6.2-9.6-14.42-9.6-23.17 0-8.75 3.4-16.98 9.6-23.17 6.2-6.2 14.42-9.6 23.17-9.6 8.75 0 16.98 3.4 23.18 9.6 12.77 12.75 12.77 33.55 0 46.33-6.2 6.2-14.43 9.6-23.18 9.6z" />
-          </svg>
+        <div
+          className="autocomplete-input__search-icon"
+          onClick={this.handleSubmit}
+        >
+          <SearchIcon />
         </div>
       </form>
     )
