@@ -1,4 +1,4 @@
-import React, {Component} from 'react';
+import React, { Component } from 'react'
 import cx from 'classnames'
 import './AutocompleteInputBox.scss'
 
@@ -6,16 +6,14 @@ class AutocompleteInputBox extends Component {
   render() {
     const { children, footer, className } = this.props
     return (
-      <div className={ cx("autocomplete-input-box", className) }>
-                {children}
+      <div className={cx('autocomplete-input-box', className)}>
+        {children}
         {footer && (
-          <div className="autocomplete-input-box__footer">
-            {footer}
-          </div>
+          <div className="autocomplete-input-box__footer">{footer}</div>
         )}
       </div>
-    );
+    )
   }
 }
 
-export default AutocompleteInputBox;
+export default AutocompleteInputBox
