@@ -1,5 +1,5 @@
-import React, {Component} from 'react';
-import DOMPurify from 'dompurify';
+import React, { Component } from 'react'
+import DOMPurify from 'dompurify'
 import './QuickStatsBar.scss'
 
 import TreeShakeIcon from '../../assets/tree-shake.svg'
@@ -11,15 +11,15 @@ import InfoIcon from '../../assets/info.svg'
 
 class QuickStatsBar extends Component {
   static defaultProps = {
-    description: ''
+    description: '',
   }
 
   getStatItemCount = () => {
     const { isTreeShakeable, hasSideEffects } = this.props
     let statItemCount = 0
 
-    if (isTreeShakeable) statItemCount += 1;
-    if (!hasSideEffects) statItemCount += 1;
+    if (isTreeShakeable) statItemCount += 1
+    if (hasSideEffects !== true) statItemCount += 1
     return statItemCount
   }
 
@@ -27,7 +27,7 @@ class QuickStatsBar extends Component {
     let trimmed
     const { description } = this.props
     if (description.trim().endsWith('.')) {
-      trimmed = description.substring(0, description.length - 1);
+      trimmed = description.substring(0, description.length - 1)
     } else {
       trimmed = description.trim()
     }
@@ -36,7 +36,13 @@ class QuickStatsBar extends Component {
   }
 
   render() {
-    const { isTreeShakeable, hasSideEffects, dependencyCount, name, repository } = this.props
+    const {
+      isTreeShakeable,
+      hasSideEffects,
+      dependencyCount,
+      name,
+      repository,
+    } = this.props
     const statItemCount = this.getStatItemCount()
 
     return (
@@ -51,7 +57,7 @@ class QuickStatsBar extends Component {
               className="quick-stats-bar__stat--description-content"
               dangerouslySetInnerHTML={{ __html: this.getTrimmedDescription() }}
               style={{
-                maxWidth: `${500 - statItemCount * 280}px`
+                maxWidth: `${500 - statItemCount * 280}px`,
               }}
             />
           )}
@@ -59,20 +65,29 @@ class QuickStatsBar extends Component {
 
         {isTreeShakeable && (
           <div className="quick-stats-bar__stat">
-            <TreeShakeIcon className="quick-stats-bar__stat-icon"/> <span>tree-shakeable</span>
+            <TreeShakeIcon className="quick-stats-bar__stat-icon" />{' '}
+            <span>tree-shakeable</span>
           </div>
         )}
 
-        {!hasSideEffects && (
+        {!(hasSideEffects === true) && (
           <div className="quick-stats-bar__stat">
-            <SideEffectIcon className="quick-stats-bar__stat-icon"/> <span>side-effect free</span>
+            <SideEffectIcon className="quick-stats-bar__stat-icon" />{' '}
+            <span>
+              {hasSideEffects.length ? 'some side-effects' : 'side-effect free'}
+            </span>
           </div>
         )}
         <div className="quick-stats-bar__stat quick-stats-bar__stat--optional">
-          <DependencyIcon className="quick-stats-bar__stat-icon"/>
+          <DependencyIcon className="quick-stats-bar__stat-icon" />
           <span>
-            {dependencyCount === 0 ? 'no dependencies' : (
-              <span>{dependencyCount} {dependencyCount > 1 ? 'dependencies' : 'dependency'}</span>
+            {dependencyCount === 0 ? (
+              'no dependencies'
+            ) : (
+              <span>
+                {dependencyCount}{' '}
+                {dependencyCount > 1 ? 'dependencies' : 'dependency'}
+              </span>
             )}
           </span>
         </div>
@@ -83,7 +98,7 @@ class QuickStatsBar extends Component {
             target="_blank"
             rel="noopener noreferrer"
           >
-            <NPMIcon className="quick-stats-bar__logo-icon quick-stats-bar__logo-icon--npm"/>
+            <NPMIcon className="quick-stats-bar__logo-icon quick-stats-bar__logo-icon--npm" />
           </a>
           <a
             className="quick-stats-bar__link"
@@ -91,13 +106,12 @@ class QuickStatsBar extends Component {
             target="_blank"
             rel="noopener noreferrer"
           >
-            <GithubIcon
-              className="quick-stats-bar__logo-icon quick-stats-bar__logo-icon quick-stats-bar__logo-icon--github"/>
+            <GithubIcon className="quick-stats-bar__logo-icon quick-stats-bar__logo-icon quick-stats-bar__logo-icon--github" />
           </a>
         </div>
       </div>
-    );
+    )
   }
 }
 
-export default QuickStatsBar;
+export default QuickStatsBar
