@@ -1,6 +1,5 @@
 require('dotenv-defaults').config()
 const next = require('next')
-const limit = require('./server/rate-limit-middleware')
 const exec = require('execa')
 
 const Koa = require('koa')
@@ -18,6 +17,9 @@ const { parsePackageString } = require('./utils/common.utils')
 const firebaseUtils = require('./utils/firebase.utils')
 const logger = require('./server/Logger')
 
+const limit = require('./server/middlewares/rateLimit.middleware')
+const exportsMiddlware = require('./server/middlewares/exports.middleware')
+const exportsSizesMiddlware = require('./server/middlewares/exportsSizes.middleware')
 const resolvePackageMiddleware = require('./server/middlewares/results/resolvePackage.middleware')
 const cachedResponseMiddleware = require('./server/middlewares/results/cachedResponse.middleware')
 const buildMiddleware = require('./server/middlewares/results/build.middleware')
@@ -26,6 +28,7 @@ const blockBlacklistMiddleware = require('./server/middlewares/results/blockBlac
 const requestLoggerMiddleware = require('./server/middlewares/requestLogger.middleware')
 const similarPackagesMiddleware = require('./server/middlewares/similar-packages/similarPackages.middleware')
 const generateImgMiddleware = require('./server/middlewares/generateImg.middleware')
+const jsonCacheMiddleware = require('./server/middlewares/jsonCache.middleware')
 
 const config = require('./server/config')
 
@@ -39,84 +42,105 @@ app.prepare().then(() => {
   const server = new Koa()
   const router = new Router()
 
-  server.use(requestId());
-  server.use(bodyParser());
+  server.use(requestId())
+  server.use(bodyParser())
   server.use(requestLoggerMiddleware)
   server.use(cacheControl())
 
   if (!dev) {
-    server.use(limit({
-      duration: 1000 * 60 * 5, //  5 mins
-      max: 60,
-    }));
+    server.use(
+      limit({
+        duration: 1000 * 60 * 5, //  5 mins
+        max: 60,
+      })
+    )
   }
 
   server.use(async (ctx, next) => {
     try {
-      await next();
+      await next()
     } catch (err) {
       if (401 == err.status) {
-        ctx.status = 401;
-        ctx.set('WWW-Authenticate', 'Basic');
-        ctx.body = 'Permission denied';
+        ctx.status = 401
+        ctx.set('WWW-Authenticate', 'Basic')
+        ctx.body = 'Permission denied'
       } else {
-        throw err;
+        throw err
       }
     }
-  });
+  })
 
-  server.use(compress({
-    filter: function (contentType) {
-      return /(text|json|javascript|svg)/.test(contentType)
-    },
-    threshold: 2048,
-    flush: require('zlib').Z_SYNC_FLUSH,
-  }))
+  server.use(
+    compress({
+      filter: function (contentType) {
+        return /(text|json|javascript|svg)/.test(contentType)
+      },
+      threshold: 2048,
+      flush: require('zlib').Z_SYNC_FLUSH,
+    })
+  )
 
-  server.use(serve('./client/assets/public', {
-    maxage: config.CACHE.PUBLIC_ASSETS * 1000,
-  }))
+  server.use(
+    serve('./client/assets/public', {
+      maxage: config.CACHE.PUBLIC_ASSETS * 1000,
+    })
+  )
 
-  server.use(proxy({
-    match: /^\/-\/search/,
-    host: 'https://www.npmjs.com',
-  }));
+  server.use(
+    proxy({
+      match: /^\/-\/search/,
+      host: 'https://www.npmjs.com',
+    })
+  )
 
-  server.use(koaCache({
-    async get(key) {
-      // Emulate koa-cash cache value
-      const value = await cache.get(key)
-      return {
-        body: value,
-        type: 'application/json',
-      }
-    },
-    set(key, value) {
-      // We only need the body part from what
-      // koa-cash gives us
-      cache.set(key, JSON.parse(value.body))
-    },
-    hash(ctx) {
-      const { name, version } = ctx.state.resolved
-      return { name, version }
-    },
-  }))
-
-  router.get('/api/size',
+  router.get(
+    '/api/size',
+    jsonCacheMiddleware({
+      get: key => cache.getPackageSize(key),
+      set: (key, value) => cache.setPackageSize(key, value),
+      hash: ctx => ({
+        name: ctx.state.resolved.name,
+        version: ctx.state.resolved.version,
+      }),
+    }),
     errorMiddleware,
     resolvePackageMiddleware,
     blockBlacklistMiddleware,
     cachedResponseMiddleware,
-    buildMiddleware,
+    buildMiddleware
   )
 
-  router.get('/api/recent', async (ctx) => {
+  router.get(
+    '/api/exports',
+    errorMiddleware,
+    resolvePackageMiddleware,
+    blockBlacklistMiddleware,
+    exportsMiddlware
+  )
+
+  router.get(
+    '/api/exports-sizes',
+    jsonCacheMiddleware({
+      get: key => cache.getExportsSize(key),
+      set: (key, value) => cache.setExportsSize(key, value),
+      hash: ctx => ({
+        name: ctx.state.resolved.name,
+        version: ctx.state.resolved.version,
+      }),
+    }),
+    errorMiddleware,
+    resolvePackageMiddleware,
+    blockBlacklistMiddleware,
+    cachedResponseMiddleware,
+    exportsSizesMiddlware
+  )
+
+  router.get('/api/recent', async ctx => {
     try {
       ctx.cacheControl = {
         maxAge: config.CACHE.RECENTS_API,
       }
-      ctx.body = await firebaseUtils
-        .getRecentSearches(ctx.query.limit)
+      ctx.body = await firebaseUtils.getRecentSearches(ctx.query.limit)
     } catch (err) {
       console.error('in /api/recent', err)
       logger.error('RECENT', err, 'RECENT FAILED: failed')
@@ -125,7 +149,7 @@ app.prepare().then(() => {
     }
   })
 
-  router.get('/api/package-history', async (ctx) => {
+  router.get('/api/package-history', async ctx => {
     const { name } = parsePackageString(ctx.query.package)
     try {
       ctx.cacheControl = {
@@ -134,7 +158,11 @@ app.prepare().then(() => {
       ctx.body = await firebaseUtils.getPackageHistory(name, ctx.query.limit)
     } catch (err) {
       console.error(err)
-      logger.error('HISTORY', err, 'HISTORY FAILED: for package' + ctx.query.package)
+      logger.error(
+        'HISTORY',
+        err,
+        'HISTORY FAILED: for package' + ctx.query.package
+      )
       ctx.status = 422
       ctx.body = { type: err.name, message: err.message }
     }
@@ -144,12 +172,13 @@ app.prepare().then(() => {
 
   router.get('/api/stats-image', generateImgMiddleware)
 
-  router.get('/admin/restart',
+  router.get(
+    '/admin/restart',
     auth({ name: 'bundlephobia', pass: process.env.BASIC_AUTH_PASSWORD }),
     async (ctx, next) => {
       try {
-        const { stdout, stderr } = await exec.shell('pm2 reload all');
-        ctx.body = 'Server restarted' + stdout;
+        const { stdout, stderr } = await exec.shell('pm2 reload all')
+        ctx.body = 'Server restarted' + stdout
       } catch (err) {
         console.error('Failed to restart', err)
         ctx.status = 500
@@ -166,18 +195,21 @@ app.prepare().then(() => {
       ctx.status = 500
       ctx.body = 'Failed to restart'
     } else {
-      const { stdout, stderr } = await exec.shell('pm2 reload all');
-      ctx.body = 'Server restarted' + stdout;
+      const { stdout, stderr } = await exec.shell('pm2 reload all')
+      ctx.body = 'Server restarted' + stdout
       console.error(stderr)
     }
   })
 
-  router.get('/admin/clear-cache',
+  router.get(
+    '/admin/clear-cache',
     auth({ name: 'bundlephobia', pass: process.env.BASIC_AUTH_PASSWORD }),
     async (ctx, next) => {
       try {
-        const { stdout } = await exec.shell('rm -rf /tmp/tmp-build/cache/_cacache /tmp/tmp-build/packages/');
-        ctx.body = 'Cache cleared' + stdout;
+        const { stdout } = await exec.shell(
+          'rm -rf /tmp/tmp-build/cache/_cacache /tmp/tmp-build/packages/'
+        )
+        ctx.body = 'Cache cleared' + stdout
       } catch (err) {
         console.error('Failed to clear cache', err)
         ctx.status = 500
@@ -197,7 +229,7 @@ app.prepare().then(() => {
   })
 
   server.use(router.routes())
-  server.listen(port, (err) => {
+  server.listen(port, err => {
     if (err) throw err
     console.log(`> Ready on http://localhost:${port}`)
   })
