@@ -2,7 +2,7 @@ import React, { Component } from 'react'
 import Analytics from 'react-ga'
 import ResultLayout from 'client/components/ResultLayout'
 import Separator from 'client/components/Separator'
-import scanBlacklist from './scanBlacklist'
+import scanBlacklist from '../../client/config/scanBlacklist'
 import Dropzone from 'react-dropzone'
 import Router from 'next/router'
 import * as semver from 'semver'
@@ -11,28 +11,27 @@ import './Scan.scss'
 export default class Scan extends Component {
   state = {
     packages: null,
-    selectedPackages: []
+    selectedPackages: [],
   }
 
   componentDidMount() {
     Analytics.pageview(window.location.pathname)
   }
 
-
-  resolveVersionFromRange = (range) => {
+  resolveVersionFromRange = range => {
     const rangeSet = new semver.Range(range).set
     return rangeSet[0][0].semver.version
   }
 
   setSelectedPackages = () => {
-    const checkedInputs =
-      this.packageSelectionContainer.querySelectorAll('input:checked')
+    const checkedInputs = this.packageSelectionContainer.querySelectorAll(
+      'input:checked'
+    )
 
-    const selectedPackages = Array.from(checkedInputs)
-      .map(({ value }) => {
-        const [name, resolvedVersion] = value.split('#')
-        return { name, resolvedVersion }
-      })
+    const selectedPackages = Array.from(checkedInputs).map(({ value }) => {
+      const [name, resolvedVersion] = value.split('#')
+      return { name, resolvedVersion }
+    })
 
     this.setState({ selectedPackages })
   }
@@ -68,14 +67,13 @@ export default class Scan extends Component {
           action: 'package.json dropped',
           value: packages.length,
         })
-
       } catch (err) {
         this.showInvalidFileError()
       }
     }
 
     try {
-      reader.readAsBinaryString(file);
+      reader.readAsBinaryString(file)
     } catch (err) {
       console.error(err)
       this.showInvalidFileError()
@@ -120,9 +118,9 @@ export default class Scan extends Component {
         <div>
           <Dropzone
             className="scan__dropzone"
-            onDropAccepted={ this.handleDropAccepted }
-            onDropRejected={ this.handleDropRejected }
-            multiple={ false }
+            onDropAccepted={this.handleDropAccepted}
+            onDropRejected={this.handleDropRejected}
+            multiple={false}
             accept="application/json"
           >
             <p>
@@ -140,48 +138,41 @@ export default class Scan extends Component {
         <div>
           <header className="scan__selection-header">
             <h1 className="scan__page-title"> Select packages to scan </h1>
-            <button className="scan__btn" onClick={ this.handleScanClick }>
-              Scan { selectedPackages.length } packages
+            <button className="scan__btn" onClick={this.handleScanClick}>
+              Scan {selectedPackages.length} packages
             </button>
-            <button className="scan__btn" onClick={ this.handleResetClick }>
+            <button className="scan__btn" onClick={this.handleResetClick}>
               Reset
             </button>
           </header>
           <ul
             className="scan__package-container"
-            ref={ pc => (this.packageSelectionContainer = pc) }
+            ref={pc => (this.packageSelectionContainer = pc)}
           >
-            {
-              packages.map(({ name, versionRange, resolvedVersion }) => (
-                <li className="scan__package-item" key={ name }>
-                  <label>
-                    <input
-                      type="checkbox"
-                      defaultChecked={
-                        !scanBlacklist.some(regex => regex.test(name))
-                      }
-                      value={ `${name}#${resolvedVersion}` }
-                      onChange={ this.handleSelectionChange }
-                    />
-                    <span className="scan__package-item-title">
-                    <span>{ name }</span>
+            {packages.map(({ name, versionRange, resolvedVersion }) => (
+              <li className="scan__package-item" key={name}>
+                <label>
+                  <input
+                    type="checkbox"
+                    defaultChecked={
+                      !scanBlacklist.some(regex => regex.test(name))
+                    }
+                    value={`${name}#${resolvedVersion}`}
+                    onChange={this.handleSelectionChange}
+                  />
+                  <span className="scan__package-item-title">
+                    <span>{name}</span>
                     <span className="scan__package-item-version">
-                    { versionRange } &rarr; { resolvedVersion }
+                      {versionRange} &rarr; {resolvedVersion}
                     </span>
                   </span>
-                  </label>
-                </li>
-              ))
-            }
+                </label>
+              </li>
+            ))}
           </ul>
         </div>
       )
     }
-    return (
-      <ResultLayout className="scan-page">
-
-        { content }
-      </ResultLayout>
-    )
+    return <ResultLayout className="scan-page">{content}</ResultLayout>
   }
 }
