@@ -1,6 +1,8 @@
 module.exports = {
- plugins: [
-  require('postcss-easy-import')({prefix: '_'}), // keep this first
-  require('autoprefixer')({ /* ...options */ }) // so imports are auto-prefixed too
- ]
+  plugins: [
+    require('postcss-easy-import')({ prefix: '_' }), // keep this first
+    require('autoprefixer')({
+      /* ...options */
+    }), // so imports are auto-prefixed too
+  ],
 }
