@@ -23,8 +23,21 @@ module.exports = {
     /cheat-\d+/,
     /-hacks?-/,
     /^nuxt$/,
-    /react-scripts/
+    /^next$/,
+    /^react-scripts/,
+    /^polymer-cli/,
+    /^parcel$/,
+    /^devextreme$/,
+    /^yarn$/,
   ],
+
+  unsupported: [
+    {
+      test: /^@types\//,
+      reason: "Type packages don't usually contain any runtime code.",
+    },
+  ],
+
   CACHE: {
     PUBLIC_ASSETS: dev ? 0 : 24 * 60 * 60,
     RECENTS_API: dev ? 0 : 20 * 60,
@@ -33,6 +46,7 @@ module.exports = {
     SIZE_API_DEFAULT: dev ? 0 : 30,
     SIZE_API_ERROR: dev ? 0 : 60,
     SIZE_API_ERROR_FATAL: dev ? 0 : 60 * 60,
-    SIZE_API_HAS_VERSION: dev ? 0 : 12 * 60 * 60,
-  }
+    SIZE_API_ERROR_UNSUPPORTED: dev ? 0 : 24 * 60 * 60,
+    SIZE_API_HAS_VERSION: dev ? 0 : 24 * 60 * 60,
+  },
 }
