@@ -1,10 +1,7 @@
+// Ununsed file
+
 import dateTimeList from './date-time'
 import markdownList from './markdown'
 import storageList from './storage'
 
-
-export default [
-  ...dateTimeList,
-  ...markdownList,
-  ...storageList,
-]
+export default [...dateTimeList, ...markdownList, ...storageList]
