@@ -1,4 +1,4 @@
-const got = require("got")
+const got = require('got')
 const remark = require('remark')
 const strip = require('strip-markdown')
 const natural = require('natural')
@@ -13,21 +13,25 @@ const CONFIG = require('../../config')
 const MIN_CUTOFF_SCORE = 12
 
 const prefixURL = (url, { base, user, project, head, path }) => {
-  if (url.indexOf('//') > 0) {
-    return url;
+  if (url.includes('//')) {
+    return url
   } else {
     return new URL(
       (path ? path.replace(/^\//, '') + '/' : '') +
-      url.replace(/^(\.?\/?)/, ''),
+        url.replace(/^(\.?\/?)/, ''),
       `${base}/${user}/${project}/${path ? '' : `${head}/`}`
-    );
+    )
   }
-};
+}
 
 async function getPackageDetails(packageName) {
   let readme
-  const { body } = await got(
-    `https://ofcncog2cu-dsn.algolia.net/1/indexes/npm-search/${encodeURIComponent(packageName)}?x-algolia-application-id=OFCNCOG2CU&x-algolia-api-key=f54e21fa3a2a0160595bb058179bfb1e`,
+  const {
+    body,
+  } = await got(
+    `https://ofcncog2cu-dsn.algolia.net/1/indexes/npm-search/${encodeURIComponent(
+      packageName
+    )}?x-algolia-application-id=OFCNCOG2CU&x-algolia-api-key=f54e21fa3a2a0160595bb058179bfb1e`,
     { json: true }
   )
 
@@ -44,13 +48,16 @@ async function getPackageDetails(packageName) {
 async function getReadme(repository) {
   const { host, user, project, branch, path } = repository
   if (host === 'github.com') {
-    const getGithubFile = async (fileName) => await got(prefixURL(fileName, {
-      base: 'https://raw.githubusercontent.com',
-      user,
-      project,
-      head: branch,
-      path: path.replace(/\/tree\//, ''),
-    }))
+    const getGithubFile = async fileName =>
+      await got(
+        prefixURL(fileName, {
+          base: 'https://raw.githubusercontent.com',
+          user,
+          project,
+          head: branch,
+          path: path.replace(/\/tree\//, ''),
+        })
+      )
 
     try {
       const { body } = await getGithubFile('README.md')
@@ -68,7 +75,7 @@ async function getReadme(repository) {
       // Once gitlab adds support, we can get rid of this workaround.
       const apiUrl = `https://gitlab.com/api/v4/projects/${user}%2F${project}/repository/files/${encodeURIComponent(
         filePath
-      )}?ref=${branch}`;
+      )}?ref=${branch}`
       const { body } = await got(apiUrl, { json: true })
 
       if (body.encoding === 'base64') {
@@ -76,7 +83,7 @@ async function getReadme(repository) {
       } else {
         return body.content
       }
-    };
+    }
 
     return getGitlabFile({
       user,
@@ -85,8 +92,11 @@ async function getReadme(repository) {
       filePath: `${path}/README.md`,
     })
   } else if (host === 'bitbucket.org') {
-
-    const { body } = await got(`https://bitbucket.org/${user}/${project}${path ? path.replace('src', 'raw') : `/raw/${branch}`}/README.md`)
+    const { body } = await got(
+      `https://bitbucket.org/${user}/${project}${
+        path ? path.replace('src', 'raw') : `/raw/${branch}`
+      }/README.md`
+    )
     return body
   }
 }
@@ -99,7 +109,8 @@ async function stripMarkdown(readme) {
         if (err) reject(err)
         resolve(
           String(file).replace(
-            /\b(npm|code|library|Node|example|project|license|MIT)\b/ig, ''
+            /\b(npm|code|library|Node|example|project|license|MIT)\b/gi,
+            ''
           )
         )
       })
@@ -118,14 +129,14 @@ function getScore(categoryTokens, packageTokens) {
 }
 
 function getInCategoryMap(packageName) {
-  return Object.keys(categories)
-    .find(label =>
-      categories[label].similar.some(similarPackage => similarPackage === packageName)
+  return Object.keys(categories).find(label =>
+    categories[label].similar.some(
+      similarPackage => similarPackage === packageName
     )
+  )
 }
 
 async function getCategory(packageName) {
-
   if (getInCategoryMap(packageName)) {
     return {
       label: getInCategoryMap(packageName),
@@ -134,16 +145,14 @@ async function getCategory(packageName) {
   }
 
   const { description, keywords } = await getPackageDetails(packageName)
-  const tokenizer = new natural.WordTokenizer();
-  const tokenString = await stripMarkdown(description) + ' ' + keywords.join(' ')
-  const packageTokens =
-    tokenizer.tokenize(tokenString)
-      .map(token => token.toLowerCase())
-      .map(natural.PorterStemmer.stem)
-      .concat(
-        tokenizer.tokenize(packageName)
-          .map(natural.PorterStemmer.stem)
-      )
+  const tokenizer = new natural.WordTokenizer()
+  const tokenString =
+    (await stripMarkdown(description)) + ' ' + keywords.join(' ')
+  const packageTokens = tokenizer
+    .tokenize(tokenString)
+    .map(token => token.toLowerCase())
+    .map(natural.PorterStemmer.stem)
+    .concat(tokenizer.tokenize(packageName).map(natural.PorterStemmer.stem))
 
   const scores = {}
   let maxScoreCategory = {
@@ -153,14 +162,12 @@ async function getCategory(packageName) {
 
   Object.keys(categories).forEach(label => {
     const categoryTokens = flatten(
-      categories[label]
-        .tags
-        .map(tagObj =>
-          tokenizer.tokenize(tagObj.tag).map(tokenizedTag => ({
-              tag: natural.PorterStemmer.stem(tokenizedTag).toLowerCase(),
-              weight: tagObj.weight
-            })
-          ))
+      categories[label].tags.map(tagObj =>
+        tokenizer.tokenize(tagObj.tag).map(tokenizedTag => ({
+          tag: natural.PorterStemmer.stem(tokenizedTag).toLowerCase(),
+          weight: tagObj.weight,
+        }))
+      )
     )
 
     const score = getScore(categoryTokens, packageTokens)
@@ -181,8 +188,17 @@ async function test() {
     categories[label].similar.forEach(async pack => {
       const actualCategory = await getCategory(pack)
 
-      if ((!actualCategory) || actualCategory.category !== label || actualCategory.score < MIN_CUTOFF_SCORE) {
-        debugTest('Package %s. Category expected: %s, got: %o', pack, label, actualCategory)
+      if (
+        !actualCategory ||
+        actualCategory.category !== label ||
+        actualCategory.score < MIN_CUTOFF_SCORE
+      ) {
+        debugTest(
+          'Package %s. Category expected: %s, got: %o',
+          pack,
+          label,
+          actualCategory
+        )
       }
     })
   })
@@ -207,7 +223,7 @@ async function similarPackagesMiddleware(ctx) {
           ...matchedCategory,
           label: value.name,
           tags: value.tags,
-          similar: value.similar.filter(pack => pack !== name)
+          similar: value.similar.filter(pack => pack !== name),
         },
       }
     } else {
@@ -216,7 +232,7 @@ async function similarPackagesMiddleware(ctx) {
         category: {
           label: null,
           score: 0,
-          similarPackages: []
+          similarPackages: [],
         },
       }
     }
@@ -227,14 +243,17 @@ async function similarPackagesMiddleware(ctx) {
       error: err,
     }
 
-    logger.error('SIMILAR_PACKAGES_ERROR', {
-      type: 'SIMIAR_PACKAGES',
-      requestId: ctx.state.id,
-      name,
-      details: err,
-    }, `SIMILAR PACKAGES FAILED: ${name}`)
+    logger.error(
+      'SIMILAR_PACKAGES_ERROR',
+      {
+        type: 'SIMIAR_PACKAGES',
+        requestId: ctx.state.id,
+        name,
+        details: err,
+      },
+      `SIMILAR PACKAGES FAILED: ${name}`
+    )
   }
 }
 
 module.exports = similarPackagesMiddleware
-
