@@ -1,1 +1,1 @@
-console.log('I\'m not a blacklisted package, hence will throw')
+console.log("I'm not a blacklisted package, hence will throw")
