@@ -1,18 +1,14 @@
 // Firebase does not accept a
 // few special characters for keys
 function encodeFirebaseKey(key) {
-  return key
-    .replace(/[.]/g, ',')
-    .replace(/\//g, '__')
+  return key.replace(/[.]/g, ',').replace(/\//g, '__')
 }
 
 function decodeFirebaseKey(key) {
-  return key
-    .replace(/[,]/g, '.')
-    .replace(/__/g, '/')
+  return key.replace(/[,]/g, '.').replace(/__/g, '/')
 }
 
-const formatSize = (value) => {
+const formatSize = value => {
   let unit, size
   if (Math.log10(value) < 3) {
     unit = 'B'
@@ -28,7 +24,7 @@ const formatSize = (value) => {
   return { unit, size }
 }
 
-const formatTime = (value) => {
+const formatTime = value => {
   let unit, size
   if (value < 0.5) {
     unit = 'ms'
@@ -41,18 +37,55 @@ const formatTime = (value) => {
   return { unit, size }
 }
 
-
 // Picked up from http://www.webpagetest.org/
 // Speed in KB/s
 
 const DownloadSpeed = {
-  TWO_G: 30,     // 2G Edge
-  THREE_G: 50    // Emerging markets 3G
+  TWO_G: 30, // 2G Edge
+  THREE_G: 50, // Emerging markets 3G
 }
-const getTimeFromSize = (sizeInBytes) => {
+const getTimeFromSize = sizeInBytes => {
   return {
     twoG: sizeInBytes / 1024 / DownloadSpeed.TWO_G,
     threeG: sizeInBytes / 1024 / DownloadSpeed.THREE_G,
+  }
+}
+
+function randomFromArray(arr) {
+  return arr[Math.floor(Math.random() * arr.length)]
+}
+
+function zeroToN(n) {
+  return Array.from(Array(n).keys())
+}
+
+function resolveBuildError(resultsError) {
+  if (!resultsError) {
+    return {
+      errorName: null,
+      errorBody: null,
+      errorDetails: null,
+    }
+  }
+  const errorName = resultsError.error
+    ? resultsError.error.code
+    : 'InternalServerError'
+  const errorBody = resultsError.error
+    ? resultsError.error.message
+    : 'Something went wrong!'
+  const errorDetails =
+    resultsError.error &&
+    resultsError.error.details &&
+    resultsError.error.details.originalError
+      ? Array.isArray(resultsError.error.details.originalError)
+        ? resultsError.error.details.originalError[0]
+        : resultsError.error.details.originalError.toString()
+      : null
+
+  return {
+    errorName,
+    errorBody,
+    errorDetails,
   }
 }
 
@@ -62,5 +95,8 @@ module.exports = {
   formatTime,
   formatSize,
   getTimeFromSize,
+  randomFromArray,
+  zeroToN,
+  resolveBuildError,
   DownloadSpeed,
 }
